requirements.txt
# Python dependencies for Synthea2OMOP-ETL

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Database connectivity
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0

# Testing and development
pytest>=6.2.5
pytest-cov>=2.12.1
flake8>=3.9.2
black>=21.6b0

# Logging and monitoring
structlog>=21.1.0

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.3.0
